requirements.txt
asgiref==3.3.0
astroid==2.4.2
certifi==2020.12.5
colorama==0.4.4
Django==3.1.2
djangorestframework==3.12.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pip==20.2.4
psycopg2==2.8.5
pylint==2.6.0
pytz==2020.1
setuptools==50.3.0.post20201006
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
wheel==0.35.1
wincertstore==0.2
wrapt==1.11.2
